I have removed a bug related to ATI driver on macOS
however, this is some issue regarding skipped/repeated frame... i know what the problem is ... has to do with READ_FRAMEBUFFER and WRITE_FRAMEBUFFER state logic not being 100% right. This needs to be fixed properly, but it is not a show stopper at the moment. @gnani-g
NOTE: i have disabled all logic related to Multisampling